unclutter server.ts
