+++ src/helpers/errors.ts
@@ -1,0 +1,7 @@
+export const serverError = {
+  error: {
+    code: 'ServerError',
+    message: 'an unexpected server error occurred.',
+    target: ''
+  }
+}
